Bugfix: Get multiple recipe endpoints
