+++ backend/controllers/tag.ts
@@ -2,6 +2,7 @@ import express from 'express';
 import { validateTag } from '../middelwares/inputValidation';
 import { isAuthorizedAdmin } from '../middelwares/jwtCheck';
 import * as tagService from '../services/tag';
+import * as recipeService from '../services/recipe';
 import { errorHandler } from '../util/errorHandler';
 
 const router = express.Router();
@@ -46,7 +47,7 @@ router.get('/:name/recipes', (req, res) => {
     let limit: number = req.query.limit ? parseInt(req.query.limit as string) : 0;
     let offset: number = req.query.offset ? parseInt(req.query.offset as string) : 0;
 
-    tagService.getRecipesByTag(req.params.name, limit, offset).then(recipes => {
+    recipeService.getRecipesByTag(req.params.name, limit, offset).then(recipes => {
         res.status(200).json(recipes);
     }).catch(err => {
         errorHandler(err, req, res);
