+++ backend/services/recipe.ts
@@ -120,46 +120,8 @@ export function getRandomRecipe(categoryId: string, tagId: string): Promise<Reci
 export function getAllRecipes(limit: number, offset: number): Promise<RecipeSmall[]> {
     return new Promise<RecipeSmall[]>((resolve, reject) => {
         recipeDAO.getRecipes(limit, offset).then(recipes => {
-            //Create promise to get additional data for each recipe
-            let convertRecipes = new Promise<RecipeSmall[]>((resolve, reject) => {
-            let allRecipes: RecipeSmall[] = [];
-                recipes.forEach(async (recipe, index, array) => {
-                    await recipeDAO.getPicturesFromRecipe(recipe.id).then(async pictures => {
-                        //If no picture return undefined
-                        let firstPicture : string = pictures[0]?.picture_id;
-                         await tagDAO.getRecipeTags(recipe.id).then(async tags => {
-                            recipe.tags = tags;
-                            //Get rating
-                            await recipeDAO.getRecipeRating(recipe.id).then(rating => {
-                                let recipeSmall: RecipeSmall = {
-                                    id: recipe.id,
-                                    name: recipe.name,
-                                    description: recipe.description,
-                                    tags: recipe.tags,
-                                    picture_uri: firstPicture,
-                                    createdAt: recipe.createdAt,
-                                    difficulty: recipe.difficulty,
-                                    time: recipe.time,
-                                    category_id: recipe.category_id,
-                                    user_id: recipe.user_id,
-                                    rating: rating.rating || 0
-                                };
-                                allRecipes.push(recipeSmall);
-                                if (index === array.length -1) resolve(allRecipes);
-                            }).catch(err => {
-                                reject(err);
-                            });
-
-                        }).catch(err => {
-                            reject(err);
-                        });
-                    }).catch(err => {
-                        reject(err);
-                    });
-                });
-            });
             //Wait for converter promise to finish
-            convertRecipes.then((allRecipes) => {
+            convertRecipeToRecipeSmall(recipes).then((allRecipes) => {
                 resolve(allRecipes);
             }).catch(err => {
                 reject(err);
@@ -173,45 +135,8 @@ export function getAllRecipes(limit: number, offset: number): Promise<RecipeSmal
 export function getRecipesFromUser(userID: string): Promise<RecipeSmall[]> {
     return new Promise<RecipeSmall[]>((resolve, reject) => {
         recipeDAO.getRecipesFromUser(userID).then(recipes => {
-            let convertRecipes = new Promise<RecipeSmall[]>((resolve, reject) => {
-                let allRecipes: RecipeSmall[] = [];
-                    recipes.forEach(async (recipe, index, array) => {
-                        await recipeDAO.getPicturesFromRecipe(recipe.id).then(async pictures => {
-                            //If no picture return undefined
-                            let firstPicture : string = pictures[0]?.picture_id;
-                             await tagDAO.getRecipeTags(recipe.id).then(async tags => {
-                                recipe.tags = tags;
-                                //Get rating
-                                await recipeDAO.getRecipeRating(recipe.id).then(rating => {
-                                    let recipeSmall: RecipeSmall = {
-                                        id: recipe.id,
-                                        name: recipe.name,
-                                        description: recipe.description,
-                                        tags: recipe.tags,
-                                        picture_uri: firstPicture,
-                                        createdAt: recipe.createdAt,
-                                        difficulty: recipe.difficulty,
-                                        time: recipe.time,
-                                        category_id: recipe.category_id,
-                                        user_id: recipe.user_id,
-                                        rating: rating.rating || 0
-                                    };
-                                    allRecipes.push(recipeSmall);
-                                    if (index === array.length -1) resolve(allRecipes);
-                                }).catch(err => {
-                                    reject(err);
-                                });
-    
-                            }).catch(err => {
-                                reject(err);
-                            });
-                        }).catch(err => {
-                            reject(err);
-                        });
-                    });
-                });
                 //Wait for converter promise to finish
-                convertRecipes.then((allRecipes) => {
+                convertRecipeToRecipeSmall(recipes).then((allRecipes) => {
                     resolve(allRecipes);
                 }).catch(err => {
                     reject(err);
@@ -225,40 +150,11 @@ export function getRecipesFromUser(userID: string): Promise<RecipeSmall[]> {
 export function getLikedRecipesFromUser(userID: string): Promise<RecipeSmall[]> {
     return new Promise<RecipeSmall[]>((resolve, reject) => {
         recipeDAO.getLikedRecipesFromUser(userID).then(recipes => {
-            let allRecipes: RecipeSmall[] = [];
-            recipes.forEach(async (recipe, index, array) => {
-                await recipeDAO.getPicturesFromRecipe(recipe.id).then(async pictures => {
-                    //If no picture return undefined
-                    let firstPicture : string = pictures[0]?.picture_id;
-                     await tagDAO.getRecipeTags(recipe.id).then(async tags => {
-                        recipe.tags = tags;
-                        //Get rating
-                        await recipeDAO.getRecipeRating(recipe.id).then(rating => {
-                            let recipeSmall: RecipeSmall = {
-                                id: recipe.id,
-                                name: recipe.name,
-                                description: recipe.description,
-                                tags: recipe.tags,
-                                picture_uri: firstPicture,
-                                createdAt: recipe.createdAt,
-                                difficulty: recipe.difficulty,
-                                time: recipe.time,
-                                category_id: recipe.category_id,
-                                user_id: recipe.user_id,
-                                rating: rating.rating || 0
-                            };
-                            allRecipes.push(recipeSmall);
-                            if (index === array.length -1) resolve(allRecipes);
-                        }).catch(err => {
-                            reject(err);
-                        });
-
-                    }).catch(err => {
-                        reject(err);
-                    });
-                }).catch(err => {
-                    reject(err);
-                });
+            //Wait for converter promise to finish
+            convertRecipeToRecipeSmall(recipes).then((allRecipes) => {
+                resolve(allRecipes);
+            }).catch(err => {
+                reject(err);
             });
         }).catch(err => {
             reject(err);
@@ -487,5 +383,75 @@ export function getRecipeRating(recipe_id: string): Promise<RatingCount> {
         }).catch((err) => {
             reject(err);
         });
+    });
+}
+
+export function getRecipesByCategory(name: string, limit: number, offset: number): Promise<RecipeSmall[]> {
+    return new Promise<RecipeSmall[]>((resolve, reject) => {
+        recipeDAO.getRecipesByCategory(name, limit, offset).then((recipes) => {
+            //Wait for converter promise to finish
+            convertRecipeToRecipeSmall(recipes).then((allRecipes) => {
+                resolve(allRecipes);
+            }).catch(err => {
+                reject(err);
+            });
+        }).catch((err) => {
+            reject(err);
+        })
+    });
+}
+
+export function getRecipesByTag(name: string, limit: number, offset: number): Promise<RecipeSmall[]> {
+    return new Promise<RecipeSmall[]>((resolve, reject) => {
+        recipeDAO.getRecipesByTag(name, limit, offset).then((recipes) => {
+            //Wait for converter promise to finish
+            convertRecipeToRecipeSmall(recipes).then((allRecipes) => {
+                resolve(allRecipes);
+            }).catch(err => {
+                reject(err);
+            });
+        }).catch((err) => {
+            reject(err);
+        })
+    });
+}
+
+function convertRecipeToRecipeSmall(recipes: Recipe[]): Promise<RecipeSmall[]> {
+    return new Promise<RecipeSmall[]>(async (resolve, reject) => {
+        let allRecipes: RecipeSmall[] = [];
+
+        for await (const recipe of recipes) {
+            await recipeDAO.getPicturesFromRecipe(recipe.id).then(async pictures => {
+                //If no picture return undefined
+                let firstPicture : string = pictures[0]?.picture_id;
+                await tagDAO.getRecipeTags(recipe.id).then(async tags => {
+                    recipe.tags = tags;
+                    //Get rating
+                    await recipeDAO.getRecipeRating(recipe.id).then(rating => {
+                        let recipeSmall: RecipeSmall = {
+                            id: recipe.id,
+                            name: recipe.name,
+                            description: recipe.description,
+                            tags: recipe.tags,
+                            picture_uri: firstPicture,
+                            createdAt: recipe.createdAt,
+                            difficulty: recipe.difficulty,
+                            time: recipe.time,
+                            category_id: recipe.category_id,
+                            user_id: recipe.user_id,
+                            rating: rating.rating || 0
+                        };
+                        allRecipes.push(recipeSmall);
+                    }).catch(err => {
+                        reject(err);
+                    });
+                }).catch(err => {
+                    reject(err);
+                });
+            }).catch(err => {
+                reject(err);
+            });
+        }
+        resolve(allRecipes);
     });
 }
